refactor data types for acs ygto mean transportation time to work pipeline

+++ acs/acs_ygt_mean_transportation_time_to_work/acs_ygt_mean_transportation_time_to_work_pipeline.py
@@ -73,8 +73,8 @@ class AcsYgtMeanTransportationTimeToWorkPipeline(EasyPipeline):
 
         dtype = {
             'year':     'UInt16',
-            'moe':      'UInt32',
-            'mea':      'UInt32',
+            'moe':      'Float32',
+            'mea':      'Float32',
             'geoid':    'String'
         }
